feat: add setData method Also rename java package and remove access details.
