+++ android/build.gradle
@@ -48,12 +48,6 @@ repositories {
         // All of React Native (JS, Obj-C sources, Android binaries) is installed from npm
         url("$rootDir/../node_modules/react-native/android")
     }
-    maven {
-      url "https://jitpack.io"
-      credentials {
-        username "jp_ek92aukm6a00g1plijctt5an8j"
-      }
-    }
     google()
     jcenter()
 }
