+++ android/src/main/AndroidManifest.xml
@@ -1,5 +1,5 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-          package="com.reactnativelandmarksid">
+          package="com.matejdr.landmarksid">
 
   <uses-permission android:name="android.permission.ACCESS_FINE_LOCATION" />
