+++ example/android/app/src/main/java/com/example/reactnativelandmarksid/MainApplication.java
@@ -10,7 +10,7 @@ import com.facebook.react.ReactInstanceManager;
 import com.facebook.soloader.SoLoader;
 import java.lang.reflect.InvocationTargetException;
 import java.util.List;
-import com.reactnativelandmarksid.LandmarksidPackage;
+import com.matejdr.landmarksid.LandmarksidPackage;
 
 public class MainApplication extends Application implements ReactApplication {
