+++ example/android/build.gradle
@@ -30,9 +30,13 @@ allprojects {
             // Android JSC is installed from npm
             url("$rootDir/../node_modules/jsc-android/dist")
         }
-
+        maven {
+          url "https://jitpack.io"
+          credentials {
+            username project.property("landmarkid_username")
+          }
+        }
         google()
         jcenter()
-        maven { url 'https://www.jitpack.io' }
     }
 }
