+++ example/android/gradle.properties
@@ -20,3 +20,4 @@
 android.useAndroidX=true
 android.enableJetifier=true
 FLIPPER_VERSION=0.54.0
+landmarkid_username="[ENTER PROVIDED USERNAME HERE]"
